+++ EchoLedger/Features/Transaction/Domain/TransactionSplit.swift
@@ -15,7 +15,7 @@ struct TransactionSplit: Identifiable, Equatable, Codable {
     // MARK: Properties
     let id: UUID
     let accountId: UUID
-    let amount: Decimal
+    let amount: Double
 
     // MARK: Init
     /// Creates a new TransactionSplit.
@@ -23,7 +23,7 @@ struct TransactionSplit: Identifiable, Equatable, Codable {
     ///   - id: Unique identifier. Defaults to a new UUID.
     ///   - accountId: The identifier of the Account this split is allocated to.
     ///   - amount: The portion of the total transaction amount for this account. Must be positive.
-    init(id: UUID = UUID(), accountId: UUID, amount: Decimal) {
+    init(id: UUID = UUID(), accountId: UUID, amount: Double) {
         self.id = id
         self.accountId = accountId
         self.amount = amount
